+++ src/pages/observation/observation.ts
@@ -1,14 +1,13 @@
 import { Component, ViewChild } from '@angular/core';
 
 import { AlertController, IonicPage, NavController, NavParams } from 'ionic-angular';
-
-import { UUID } from 'angular2-uuid';
 import * as moment from 'moment/moment';
 
 import { Dataset } from '../../biosys-core/interfaces/api.interfaces';
 import { StorageService } from '../../shared/services/storage.service';
 import { ClientRecord } from '../../shared/interfaces/mobile.interfaces';
 import { RecordFormComponent } from '../../components/record-form/record-form';
+import { APIService } from '../../biosys-core/services/api.service';
 import { PhotoGalleryComponent } from '../../components/photo-gallery/photo-gallery';
 import { from } from 'rxjs/observable/from';
 import { mergeMap } from 'rxjs/operators';
@@ -35,30 +34,34 @@ export class ObservationPage {
     private recordClientId: string;
 
     constructor(private navCtrl: NavController, private navParams: NavParams, private storageService: StorageService,
-                private alertController: AlertController) {
-    }
-
-    public ionViewDidEnter() {
-        this.recordClientId = this.navParams.get('recordClientId');
-        this.isNewRecord = !this.recordClientId;
-        if (this.isNewRecord) {
-            this.recordClientId = UUID.UUID();
+                private alertController: AlertController, private apiService: APIService) {
+        if (!this.navParams.get('datasetName')) {
+            this.showLeavingAlertMessage = false;
+            this.navCtrl.pop();
         }
-        this.photoGallery.RecordId = this.recordClientId;
-        this.parentId = this.navParams.get('parentId');
+
+        if (this.navParams.data.hasOwnProperty('parentId')) {
+            this.parentId = this.navParams.get('parentId');
+        }
+
+        const recordClientId = this.navParams.get('recordClientId');
+        this.isNewRecord = !recordClientId;
 
         this.storageService.getDataset(this.navParams.get('datasetName')).subscribe((dataset: Dataset) => {
-            this.dataset = dataset;
+            if (dataset) {
+                this.dataset = dataset;
 
-            if (!this.isNewRecord) {
-                // if this is an existing record, set form values from data
-                this.storageService.getRecord(this.recordClientId).subscribe(
-                    record => {
-                        this.record = record;
-                        this.recordForm.value = record.data;
-                        this.photoGallery.PhotoIds = record.photoIds;
-                    }
-                );
+                if (recordClientId) {
+                    // if this is an existing record, set form values from data
+                    this.storageService.getRecord(recordClientId).subscribe(
+                        record => {
+                            this.record = record;
+                            this.recordForm.value = record.data;
+                            this.photoGallery.RecordId = recordClientId;
+                            this.photoGallery.PhotoIds = record.photoIds;
+                        }
+                    );
+                }
             }
         });
     }
@@ -120,16 +123,33 @@ export class ObservationPage {
                 {
                     text: 'Yes',
                     handler: () => {
-                        this.photoGallery.rollback();
-                        this.storageService.deleteRecord(this.record.client_id).subscribe( deleted => {
-                            if (this.record.photoIds) {
-                                from(this.record.photoIds).pipe(
-                                    mergeMap( photoId => this.storageService.deletePhoto(photoId) )
-                                ).subscribe();
-                            }
+                        if (this.record) {
+                            this.photoGallery.rollback();
+                            this.storageService.deleteRecord(this.record.client_id).subscribe(deleted => {
+                                if (this.record.photoIds) {
+                                    from(this.record.photoIds).pipe(
+                                        mergeMap(photoId => this.storageService.deletePhoto(photoId))
+                                    ).subscribe();
+                                }
+                                this.showLeavingAlertMessage = false;
+                                this.navCtrl.pop();
+                            }, (error) => {
+                                this.alertController.create({
+                                    title: 'Cannot Delete',
+                                    message: 'Sorry, cannot delete this observation.',
+                                    enableBackdropDismiss: true,
+                                    buttons: [
+                                        {
+                                            text: 'OK',
+                                            handler: () => {}
+                                        }
+                                    ]
+                                }).present();
+                            });
+                        } else {
                             this.showLeavingAlertMessage = false;
                             this.navCtrl.pop();
-                        });
+                        }
                     }
                 },
                 {
@@ -137,5 +157,4 @@ export class ObservationPage {
                 }]
         }).present();
     }
-
 }
